Revert "Implémentation finale idée Hayk"
This reverts commit e30e847ec31b3df4598b557f42a62fae98c87676.

+++ src/main/java/insa/aubin/devisbatiment/modele/Point.java
@@ -4,29 +4,29 @@ import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
 public class Point implements Dessin {
-    private float x;
-    private float y;
+    private double x;
+    private double y;
     public static double RAYON_IN_DRAW = 5;
     private Color color = Color.BLACK;
     
-    public Point(float x, float y){
+    public Point(double x, double y){
         this.x = x;
         this.y = y;
     }
 
-    public float getX() {
+    public double getX() {
         return x;
     }
 
-    public void setX(float x) {
+    public void setX(double x) {
         this.x = x;
     }
 
-    public float getY() {
+    public double getY() {
         return y;
     }
 
-    public void setY(float y) {
+    public void setY(double y) {
         this.y = y;
     }
